This update solve the problem of lazyload.
On Last Sat I assumed this problem was related with inappropriate style
of using session. But it is not.

The reason for error:org.hibernate.LazyInitializationException: could
not initialize proxy - no Session is beacuse when we use lazy load the
seesion must keep open until we finish using of lazy load part. In
spring a filter -- OpenSessionInViewFilter -- is created for this. 

The reason I met the exception last week is I use the wrong order of
filter declaration in web.xml. It should declare spring configuration
then struts configuration. 

you can check the difference between this version and last version in
web.xml file